+++ src/runtime.cc
@@ -60,7 +60,7 @@ void RuntimeCollectGarbage(Heap* heap, char* stack_top) {
 }
 
 
-off_t RuntimeGetHash(Heap* heap, char* value) {
+intptr_t RuntimeGetHash(Heap* heap, char* value) {
   Heap::HeapTag tag = HValue::GetTag(value);
 
   switch (tag) {
@@ -70,7 +70,7 @@ off_t RuntimeGetHash(Heap* heap, char* value) {
    case Heap::kTagObject:
    case Heap::kTagArray:
    case Heap::kTagCData:
-    return reinterpret_cast<off_t>(value);
+    return reinterpret_cast<intptr_t>(value);
    case Heap::kTagNil:
     return 0;
    case Heap::kTagBoolean:
@@ -100,10 +100,10 @@ off_t RuntimeGetHash(Heap* heap, char* value) {
 }
 
 
-off_t RuntimeLookupProperty(Heap* heap,
+intptr_t RuntimeLookupProperty(Heap* heap,
                             char* obj,
                             char* key,
-                            off_t insert) {
+                            intptr_t insert) {
   assert(!HValue::Cast(obj)->IsGCMarked());
   assert(!HValue::Cast(obj)->IsSoftGCMarked());
 
@@ -607,8 +607,8 @@ char* RuntimeBinOp(Heap* heap, char* lhs, char* rhs) {
        case BinOp::kShr: result = lval >> rval; break;
        case BinOp::kUShr:
         {
-          off_t ulval = lval;
-          off_t urval = rval;
+          intptr_t ulval = lval;
+          intptr_t urval = rval;
           result = type == ulval >> urval;
         }
         break;
@@ -697,14 +697,14 @@ char* RuntimeCloneObject(Heap* heap, char* obj) {
       ((source_map->size() << 1) + 1) * HValue::kPointerSize);
 
   // Set mask
-  *reinterpret_cast<off_t*>(result + HObject::kMaskOffset) =
+  *reinterpret_cast<intptr_t*>(result + HObject::kMaskOffset) =
       (source_map->size() - 1) * HValue::kPointerSize;
 
   // Set map
   *reinterpret_cast<char**>(result + HObject::kMapOffset) = map;
 
   // Set map's size
-  *reinterpret_cast<off_t*>(map + HMap::kSizeOffset) = source_map->size();
+  *reinterpret_cast<intptr_t*>(map + HMap::kSizeOffset) = source_map->size();
 
   // Nullify all map's slots (both keys and values)
   uint32_t size = (source_map->size() << 1) * HValue::kPointerSize;
@@ -715,17 +715,17 @@ char* RuntimeCloneObject(Heap* heap, char* obj) {
 
 
 void RuntimeDeleteProperty(Heap* heap, char* obj, char* property) {
-  off_t offset = RuntimeLookupProperty(heap, obj, property, 0);
+  intptr_t offset = RuntimeLookupProperty(heap, obj, property, 0);
 
   // Dense arrays doesn't have keys
   if (HValue::GetTag(obj) != Heap::kTagArray || !HArray::IsDense(obj)) {
     // Nil property
-    off_t keyoffset = offset - HObject::Mask(obj) - HValue::kPointerSize;
-    *reinterpret_cast<off_t*>(HObject::Map(obj) + keyoffset) = Heap::kTagNil;
+    intptr_t keyoffset = offset - HObject::Mask(obj) - HValue::kPointerSize;
+    *reinterpret_cast<intptr_t*>(HObject::Map(obj) + keyoffset) = Heap::kTagNil;
   }
 
   // Nil value
-  *reinterpret_cast<off_t*>(HObject::Map(obj) + offset) = Heap::kTagNil;
+  *reinterpret_cast<intptr_t*>(HObject::Map(obj) + offset) = Heap::kTagNil;
 }
 
 
@@ -780,7 +780,7 @@ char* RuntimeStackTrace(Heap* heap, char** frame, char* ip) {
 
     // Detect frame enter
     if (frame != NULL &&
-        static_cast<uint32_t>(reinterpret_cast<off_t>(*(frame + 2))) ==
+        static_cast<uint32_t>(reinterpret_cast<intptr_t>(*(frame + 2))) ==
             Heap::kEnterFrameTag) {
       frame = reinterpret_cast<char**>(*(frame + 4));
     }
